+++ WeakAuras/Init.lua
@@ -8,6 +8,11 @@ if versionString == "@project-version@" then
 end
 --@end-debug@
 WeakAuras.versionString = versionString
+WeakAuras.printPrefix = "|cff9900FF".."WeakAuras: "..FONT_COLOR_CODE_CLOSE
+
+WeakAuras.prettyPrint = function(msg)
+  print(WeakAuras.printPrefix .. msg)
+end
 
 WeakAuras.PowerAurasPath = "Interface\\Addons\\WeakAuras\\PowerAurasMedia\\Auras\\"
 WeakAuras.PowerAurasSoundPath = "Interface\\Addons\\WeakAuras\\PowerAurasMedia\\Sounds\\"
@@ -17,4 +22,13 @@ function WeakAuras.RegisterRegionType()
 end
 
 function WeakAuras.RegisterRegionOptions()
+end
+
+function WeakAuras.StartProfileSystem()
+end
+function WeakAuras.StartProfileAura()
+end
+function WeakAuras.StopProfileSystem()
+end
+function WeakAuras.StopProfileAura()
 end