+++ WeakAuras/RegionTypes/Model.lua
@@ -136,9 +136,11 @@ local function modify(parent, region, data)
       model:RegisterEvent("PLAYER_FOCUS_CHANGED");
     end
     model:SetScript("OnEvent", function(self, event, unitId)
+      WeakAuras.StartProfileSystem("model");
       if (event ~= "UNIT_MODEL_CHANGED" or UnitIsUnit(unitId, data.model_path)) then
         model:SetUnit(data.model_path);
       end
+      WeakAuras.StopProfileSystem("model");
     end
     );
   else
@@ -176,8 +178,10 @@ local function modify(parent, region, data)
   if(data.advance) then
     local elapsed = 0;
     model:SetScript("OnUpdate", function(self, elaps)
+      WeakAuras.StartProfileSystem("model");
       elapsed = elapsed + (elaps * 1000);
       model:SetSequenceTime(data.sequence, elapsed);
+      WeakAuras.StopProfileSystem("model");
     end);
   else
     model:SetScript("OnUpdate", nil);
@@ -265,19 +269,23 @@ WeakAuras.RegisterRegionType("model", create, modify, default, GetProperties);
 -- Work around for movies and world map hiding all models
 do
   local function preShowModels(self, event)
+    WeakAuras.StartProfileSystem("model");
     if (event == "PLAYER_LOGIN") then
       C_Timer.After(2, preShowModels);
       return;
     end
 
     for id, isLoaded in pairs(WeakAuras.loaded) do
+      WeakAuras.StartProfileAura(id);
       if (isLoaded) then
         local data = WeakAuras.regions[id];
         if (data.regionType == "model") then
           data.region:PreShow();
         end
       end
+      WeakAuras.StopProfileAura(id);
     end
+    WeakAuras.StopProfileSystem("model");
   end
 
   local movieWatchFrame;