+++ CHANGELOG.md
@@ -1,10 +1,23 @@
 # WeakAuras 2
 
-## [8.0-beta3-1-g40e68ad](https://github.com/WeakAuras/WeakAuras2/tree/40e68adfe6b2167196b2754d5dc00e91a30d306c) (2018-04-27)
+## [8.0-beta4-12-g61c14cb](https://github.com/WeakAuras/WeakAuras2/tree/61c14cbbaf0432683da9fbc031c00832cf59042b) (2018-05-04)
 
-[Full Changelog](https://github.com/WeakAuras/WeakAuras2/compare/8.0-beta3...40e68adfe6b2167196b2754d5dc00e91a30d306c)
+[Full Changelog](https://github.com/WeakAuras/WeakAuras2/compare/8.0-beta4...61c14cbbaf0432683da9fbc031c00832cf59042b)
 
-Benjamin Staneck (1):
+Benjamin Staneck (4):
 
-- update changelog
+- add a prettyPrint function and use it in a few places
+- add names to the profiling keybinds
+- add prints to pstart and pstop
+- Update changelog
+
+Infus (7):
+
+- Fix a few bugs, remove TODOs
+- Don't early return if before potentially enabling rune watching
+- Fix Overlay colors not resetting after Condition
+- Fix issues with profiling and some animations
+- Add a "loadFunc" to Prototypes which is called on loading a aura
+- Add Profiling Support to WeakAuras
+- Merge RegionOptions on multi selection